c++: make some cp_trait_kind switch statements exhaustive
This replaces the unreachable default case in some cp_trait_kind
switches with an exhaustive listing of the _unexpected_ trait codes,
so that when adding a new trait we'll get a -Wswitch diagnostic if we
forget to handle the trait code in one of these switches.

Bootstrappend and regtested on x86_64-pc-linux-gnu, does this look OK
for trunk?

gcc/cp/ChangeLog:

	* semantics.cc (trait_expr_value): Make cp_trait_kind switch
	exhaustive.
	(finish_trait_expr): Likewise.
	(finish_trait_type): Likewise.