Add op1_range for __builtin_signbit.
This is the op1_range range-op entry for __builtin_signbit.  It allows
us to wind back through a call to signbit.

For example, on the true side of if (__builtin_signbit(x_5) != 0) we
can crop down the range of x_5 to:

	[frange] float [-Inf, -0.0 (-0x0.0p+0)] -NAN

Similarly on the false side, we can crop to:

	[frange] float [0.0 (0x0.0p+0), +Inf] +NAN

This brings about an interesting question, can we fold the second
conditional here as always false?

	void foo(float x)
	{
	  if (__builtin_signbit (x))
	    {
	      if (x > 0.0)
	        link_error();
	    }
	}

The only values for x at the second conditional are negative values,
or -NAN, so it will always evaluate to false.  ISTM that we
*shouldn't* fold this conditional as there is user observable behavior
if there is a signaling NAN.  For that matter, that is exactly what we
do in ranger-ops.  We leave the conditional in place, but ranger
is able to determine that "x" is UNDEFINED on the path leading up to
link_error:

=========== BB 3 ============
Imports: x_3(D)
Exports: x_3(D)
x_3(D)	[frange] float [-Inf, -0.0 (-0x0.0p+0)] -NAN
    <bb 3> :
    if (x_3(D) > 0.0)
      goto <bb 4>; [INV]
    else
      goto <bb 5>; [INV]

3->4  (T) x_3(D) : 	[frange] UNDEFINED
3->5  (F) x_3(D) : 	[frange] float [-Inf, -0.0 (-0x0.0p+0)] -NAN

This would allow users of ranger to query x_3 on the 3->4 and notice
it's unreachable, without VRP removing the conditional.

I believe this is correct, but would like confirmation from the experts.

gcc/ChangeLog:

	* gimple-range-op.cc: Add op1_range entry for __builtin_signbit.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/vrp-float-signbit-3.c: New test.