c-family: ICE with [[gnu::nocf_check]] [PR106937]
On Thu, Oct 06, 2022 at 05:42:41PM -0400, Jason Merrill wrote:
> On 10/4/22 19:06, Marek Polacek wrote:
> > On Fri, Sep 30, 2022 at 09:12:24AM -0400, Jason Merrill wrote:
> > > On 9/29/22 18:49, Marek Polacek wrote:
> > > > When getting the name of an attribute, we ought to use
> > > > get_attribute_name, which handles both [[ ]] and __attribute__(())
> > > > forms.  Failure to do so may result in an ICE, like here.
> > > >
> > > > Bootstrapped/regtested on x86_64-pc-linux-gnu, ok for trunk?
> > >
> > > How do we print the attributes with this patch?  Don't we also want to print
> > > the namespace, and use [[]] in the output?
> >
> > Good point, however: while the testcase indeed has an attribute
> > in the [[]] form in the typedef, here we're printing its "aka":
> >
> > warning: initialization of 'FuncPointerWithNoCfCheck' {aka 'void (__attribute__((nocf_check)) *)(void)'} from incompatible pointer type 'FuncPointer' {aka 'void (*)(void)'}
> >
> > c-pretty-print.cc doesn't seem to know how to print an [[]] attribute.
> > I could do that, but then we'd print
> >
> >    aka 'void ([[nocf_check]] *)(void)'
> >
> > in the above, but that's invalid syntax!
>
> Indeed, it should be
>
> void (* [[gnu::nocf_check]])(void)

Ok, let me fix that too, then.  I've updated the description to show
what we print with the patch, and what was printed before.

And I also noticed another thing: type_hash_canon_hash hashes only
get_attribute_name, so build_type_attribute_qual_variant will use the
same function_type tree for these two pointers:

__attribute__((nocf_check)) typedef void (*fp1)();
[[gnu::nocf_check]] typedef void (*fp2)();

so if gcc prints a diagnostic with fp2, it will show the GNU form of
the attribute, because the already existing type of fp1 will be used!
I'm not addressing it here, just an oddity I discovered.

Bootstrapped/regtested on x86_64-pc-linux-gnu, ok for trunk?

-- >8 --
When getting the name of an attribute, we ought to use
get_attribute_name, which handles both [[]] and __attribute__(())
forms.  Failure to do so may result in an ICE, like here.

pp_c_attributes_display wasn't able to print the [[]] form of
attributes, so this patch teaches it to.

When printing a pointer to function with an attribute, the attribute
should be printed after the *, because that's what it appertains to.
With this patch we print:

  aka 'void (*__attribute__((nocf_check)))()'
  aka 'void (*[[gnu::nocf_check]])()'

rather than

  aka 'void (__attribute__((nocf_check)) *)()

or crashing in the second case.

pp_c_attributes has been unused since its introduction in r56273 so
this patch removes it.

	PR c++/106937

gcc/c-family/ChangeLog:

	* c-pretty-print.cc (pp_c_specifier_qualifier_list): When
	printing a pointer to function with an attribute, print the
	attribute after the *.
	(pp_c_attributes): Remove.
	(pp_c_attributes_display): Print the [[]] form if appropriate.
	Use get_attribute_name.  Don't print a trailing space.
	* c-pretty-print.h (pp_c_attributes): Remove.

gcc/cp/ChangeLog:

	* error.cc (dump_type_prefix): When printing a pointer/reference to
	function with an attribute, print the attribute after the */&.

gcc/testsuite/ChangeLog:

	* g++.dg/diagnostic/aka5a.C: Adjust.
	* g++.dg/diagnostic/aka5b.C: Likewise.
	* gcc.dg/diag-aka-5a.c: Likewise.
	* gcc.dg/diag-aka-5b.c: Likewise.
	* c-c++-common/pointer-to-fn1.c: New test.