PR middle-end/88345: Honor -falign-functions=N even optimized for size.
Currnetly setting of -falign-functions=N will be ignored if the function
is optimized for size or marked as cold function.

However function alignment requirement is needed even optimized for
size in some situations, RISC-V target is an example, RISC-V kernel implement
patchable function that require function must be align to at least 4 bytes for
atomicly patch the function prologue.

Here is 4 way to fix that:
1. Fix -falign-functions=N, let it work as expect on -Os and all cold
functions, which is this patch.
2. Force align to 4 byte if -fpatchable-function-entry is given by adjust
RISC-V's FUNCTION_BOUNDARY.
3. Adjust RISC-V's FUNCTION_BOUNDARY to let it honor to -falign-functions=N.
4. Adding a -malign-functions=N for RISC-V...which x86 already deprecated that.

And this patch is the first approach.

gcc/ChangeLog:

	PR middle-end/88345
	* varasm.cc (assemble_start_function): Adjust function align
	even optimized for size.
	* doc/invoke.texi (Os): Remove -falign-functions= from the exclusion
	list of -Os.

gcc/testsuite/ChangeLog:

	PR middle-end/88345
	* gcc.target/i386/pr88345-1.c: New.
	* gcc.target/i386/pr88345-2.c: Ditto.
	* gcc.target/riscv/pr88345-1.c: Ditto.
	* gcc.target/riscv/pr88345-2.c: Ditto.