OpenMP: Pointers and member mappings
Implementing the "omp declare mapper" functionality, I noticed some
cases where handling of derived type members that are pointers doesn't
seem to be quite right. At present, a type such as this:

  type T
  integer, pointer, dimension(:) :: arrptr
  end type T

  type(T) :: tvar
  [...]
  !$omp target map(tofrom: tvar%arrptr)

will be mapped using three mapping nodes:

  GOMP_MAP_TO             tvar%arrptr       (the descriptor)
  GOMP_MAP_TOFROM         *tvar%arrptr%data (the actual array data)
  GOMP_MAP_ALWAYS_POINTER tvar%arrptr%data  (a pointer to the array data)

This follows OMP 5.0, 2.19.7.1 (or OpenMP 5.2, 5.8.3) "map Clause":

  "If a list item in a map clause is an associated pointer and the
   pointer is not the base pointer of another list item in a map clause
   on the same construct, then it is treated as if its pointer target
   is implicitly mapped in the same clause. For the purposes of the map
   clause, the mapped pointer target is treated as if its base pointer
   is the associated pointer."

However, we can also write this:

  map(to: tvar%arrptr) map(tofrom: tvar%arrptr(3:8))

and then instead we should follow (OpenMP 5.2, 5.8.3 "map Clause"):

  "For map clauses on map-entering constructs, if any list item has a base
   pointer for which a corresponding pointer exists in the data environment
   upon entry to the region and either a new list item or the corresponding
   pointer is created in the device data environment on entry to the region,
   then:
   1. [Fortran] The corresponding pointer variable is associated with
      a pointer target that has the same rank and bounds as the pointer
      target of the original pointer, such that the corresponding list item
      can be accessed through the pointer in a target region.
   2. The corresponding pointer variable becomes an attached pointer
      for the corresponding list item."

But, that's not implemented quite right at the moment (and completely
breaks once we introduce declare mappers), because we still map the "to:
tvar%arrptr" as the descriptor and the entire array, then we map the
"tvar%arrptr(3:8)" part using the descriptor (again!) and the array
slice. The solution is to detect when we're mapping a smaller part of
the array (or a subcomponent) on the same directive, and only map the
descriptor in that case. So we get mappings like this instead:

  map(to: tvar%arrptr)   -->
  GOMP_MAP_ALLOC  tvar%arrptr  (the descriptor)

  map(tofrom: tvar%arrptr(3:8)   -->
  GOMP_MAP_TOFROM tvar%arrptr%data(3) (size 8-3+1, etc.)
  GOMP_MAP_ALWAYS_POINTER tvar%arrptr%data (bias 3, etc.)

This version of the patch alters the way that expressions are compared,
so that e.g. "tvar(i)%arrptr" is not considered to overlap/reference the
same array as "tvar(j)%arrptr(3:8)".  This is done using a new function,
gfc_omp_expr_prefix_same, which only considers element accesses equal if
they can be proven the same, i.e. if they use the same index variable.
Together with the "unordered struct" patch at the end of this series,
that allows certain types of mapping to work that don't at present.

2022-10-09  Julian Brown  <julian@codesourcery.com>

gcc/fortran/
	* dependency.cc (gfc_omp_expr_prefix_same): New function.
	* dependency.h (gfc_omp_expr_prefix_same): Add prototype.
        * gfortran.h (gfc_omp_namelist): Add "duplicate_of" field to "u2"
        union.
        * trans-openmp.cc (dependency.h): Include.
        (gfc_trans_omp_array_section): Do not map descriptors here for OpenMP.
        (gfc_symbol_rooted_namelist): New function.
        (gfc_trans_omp_clauses): Check subcomponent and subarray/element
        accesses elsewhere in the clause list for pointers to derived types or
        array descriptors, and map just the pointer/descriptor if we have any.

libgomp/
        * testsuite/libgomp.fortran/map-subarray.f90: New test.
        * testsuite/libgomp.fortran/map-subarray-2.f90: New test.
	* testsuite/libgomp.fortran/map-subarray-4.f90: New test.
	* testsuite/libgomp.fortran/map-subarray-6.f90: New test.
        * testsuite/libgomp.fortran/map-subcomponents.f90: New test.
        * testsuite/libgomp.fortran/struct-elem-map-1.f90: Adjust for
        descriptor-mapping changes.